+++ code/modules/mob/living/bot/farmbot.dm
@@ -372,38 +372,34 @@
 /obj/item/farmbot_arm_assembly/attackby(obj/item/W as obj, mob/user as mob)
 	..()
 	if((istype(W, /obj/item/analyzer/plant_analyzer)) && (build_step == 0))
+		if(!user.attempt_consume_item_for_construction(W))
+			return
 		build_step++
 		to_chat(user, "You add the plant analyzer to [src].")
 		name = "farmbot assembly"
 
-		user.remove_from_mob(W)
-		qdel(W)
-
 	else if((istype(W, /obj/item/reagent_containers/glass/bucket)) && (build_step == 1))
+		if(!user.attempt_consume_item_for_construction(W))
+			return
 		build_step++
 		to_chat(user, "You add a bucket to [src].")
 		name = "farmbot assembly with bucket"
 
-		user.remove_from_mob(W)
-		qdel(W)
-
 	else if((istype(W, /obj/item/material/minihoe)) && (build_step == 2))
+		if(!user.attempt_consume_item_for_construction(W))
+			return
 		build_step++
 		to_chat(user, "You add a minihoe to [src].")
 		name = "farmbot assembly with bucket and minihoe"
 
-		user.remove_from_mob(W)
-		qdel(W)
-
 	else if((isprox(W)) && (build_step == 3))
+		if(!user.attempt_consume_item_for_construction(W))
+			return
 		build_step++
 		to_chat(user, "You complete the Farmbot! Beep boop.")
 
 		var/mob/living/bot/farmbot/S = new /mob/living/bot/farmbot(get_turf(src), tank)
 		S.name = created_name
-
-		user.remove_from_mob(W)
-		qdel(W)
 		qdel(src)
 
 	else if(istype(W, /obj/item/pen))
